fixed low power issue and cleaned code
if advertising the device will enter low power mode in 15 seconds, this
mode has been fixed to actually go down to 1-2mA. Writing or reading
from the bosch sensor after putting it in low power mode will return it
to normal power mode.

Of course movement also returns it to normal power mode.

Additionally, the FRAM must be initialized so that it does not pull 4mA.

ADXL is initialized into low power mode immediately.

***** Known issues******
The device will stop advertising after 1 hour (advertising timeout
interval) and go into sleep mode. Something is waking the device up from
sleep mode and resetting the device. Documentation says only a interrupt
type action can do this. After this soft reset the device does not pull
the same current as it did on initial boot sequence. The device does not
reset if it is connected to a central ble device (phone)